+++ packages/animotion/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@animotion/core",
-	"version": "0.1.7",
+	"version": "0.1.8",
 	"description": "Core components for Animotion",
 	"author": "animotionjs",
 	"homepage": "https://animotion.pages.dev/",
