+++ packages/animotion/src/app.d.ts
@@ -1,12 +1,20 @@
-// See https://kit.svelte.dev/docs/types#app
-// for information about these interfaces
 declare global {
 	namespace App {
-		// interface Error {}
-		// interface Locals {}
-		// interface PageData {}
-		// interface PageState {}
-		// interface Platform {}
+		interface ViewTransition {
+			updateCallbackDone: Promise<void>
+			ready: Promise<void>
+			finished: Promise<void>
+			skipTransition: () => void
+		}
+
+		interface Document {
+			viewTransition: ViewTransition
+			startViewTransition(updateCallback: () => Promise<void> | void): ViewTransition
+		}
+
+		interface CSSStyleDeclaration {
+			viewTransitionName?: string
+		}
 	}
 }
