+++ packages/animotion/src/lib/components/index.js
@@ -1,11 +1,11 @@
-import Code from './code.svelte'
+import CodeBlock from './code-block.svelte'
 import Embed from './embed.svelte'
 import FitText from './fit.svelte'
 import Markdown from './markdown.svelte'
 import Media from './media.svelte'
 import Notes from './notes.svelte'
 import Presentation from './presentation.svelte'
-import Shiki from './shiki.svelte'
+import Code from './code.svelte'
 import Slide from './slide.svelte'
 import Stack from './stack.svelte'
 import Step from './step.svelte'
@@ -14,13 +14,13 @@ import Vertical from './vertical.svelte'
 
 export {
 	Code,
+	CodeBlock,
 	Embed,
 	FitText,
 	Markdown,
 	Media,
 	Notes,
 	Presentation,
-	Shiki,
 	Slide,
 	Stack,
 	Step,
