+++ packages/animotion/src/lib/index.js
@@ -1,1 +1,2 @@
 export * from './components/index.js'
+export * from './transition/index.js'
