+++ packages/animotion/src/lib/styles/theme.css
@@ -1,13 +1,3 @@
-:root {
-	--r-background-color: hsl(226 19% 13%);
-	--r-main-font: 'Manrope Variable';
-	--r-main-font-size: 42px;
-	--r-main-color: hsl(226 19% 98%);
-	--r-heading-font: 'Manrope';
-	--r-code-font: 'JetBrains Mono Variable';
-	--r-link-color: hsl(180 100% 50%);
-}
-
 .reveal-viewport {
 	background-color: var(--r-background-color);
 }
