+++ packages/animotion/src/routes/+page.svelte
@@ -1,74 +1,117 @@
 <script lang="ts">
-	import { Presentation, Slide, Code } from '$lib/components/index.js'
-	import { signal } from '@animotion/motion'
+	import { Presentation, Slide, Code, transition, action } from '$lib/index.js'
+	import type Reveal from 'reveal.js'
 
-	const circle = signal({ x: 0, y: 200, r: 80, fill: '#00ffff' }, { duration: 2 })
+	const options: Reveal.Options = {
+		display: 'grid',
+		disableLayout: true,
+		transition: 'slide',
+		controls: true,
+		progress: true,
+		hash: false
+	}
+
+	let text: HTMLParagraphElement
+	let code: Code
+	let boxes = $state([1, 2, 3, 4])
 </script>
 
-<Presentation>
-	<Slide animate>
-		<p class="text-8xl font-bold">🪄 Animotion</p>
-	</Slide>
+<!--
+	`use:transition` and `use:action` are just steps
+	so you can use arrow keys to step through	the slide:
 
-	<Slide animate>
-		<p class="mt-8 text-6xl font-bold">🪄 Animotion</p>
+	`transition` animates the layout and accepts:
+		- `action` (optional) callback
+		- `name` (optional)	for the transition
+		- `order` (optional) option	to specify the step order
+		- `transition` (optional)	for custom CSS transitions
+	
+	`action` is just a step like `transition`:
+		- doesn't animate the layout
+		- mostly exists as `transition` alternative but	uses
+			absolute positioning to not affect the layout
+-->
 
-		<div class="mx-auto mt-16 w-max">
-			<Code lang="ts" lines="1,4|2|3|1-4">
-				{`
-					const circle = signal(
-						{ x: 0, y: 200, r: 80, fill: '#00ffff' },
-						{ duration: 2000 }
-					)
-				`}
-			</Code>
+<Presentation {options}>
+	<Slide class="h-full place-content-center place-items-center">
+		<div use:transition>
+			<p bind:this={text} class="text-8xl font-bold drop-shadow-sm">🪄 Animotion</p>
 		</div>
-	</Slide>
 
-	<Slide on:out={() => circle.reset()} animate>
-		<p class="mt-8 text-6xl font-bold">🪄 Animotion</p>
-
-		<div class="mx-auto mt-16 w-max">
+		<div
+			use:transition={{
+				action: () => {
+					// change the DOM and animate the layout
+					text.classList.replace('text-8xl', 'text-6xl')
+				}
+			}}
+			class="mt-16"
+		>
 			<Code
-				lang="ts"
-				lines="1,4|2|3|1-4"
-				steps={{
-					'2': async () => await circle.to({ x: 400, fill: '#ffff00' }),
-					'3': async () => await circle.to({ x: 0, fill: '#00ffff' }),
-					'1-4': () => circle.reset()
-				}}
-			>
-				{`
-					async function animate() {
-						await circle.to({ x: 400, fill: '#ffff00' })
-						await circle.to({ x: 0, fill: '#00ffff' })
-					}
+				bind:this={code}
+				lang="svelte"
+				theme="poimandres"
+				code={`
+					<script>
+						let count = 0
+						let double = count * 2
+					<\/script>
+
+					<button on:click={() => count++}>
+						{double}
+					</button>
 				`}
-			</Code>
+				options={{ duration: 1, stagger: 0.3, containerStyle: false }}
+			/>
 		</div>
 
-		<svg class="mx-auto h-[400px] w-full" viewBox="0 0 400 400" data-id="animation">
-			<circle cx={$circle.x} cy={$circle.y} r={$circle.r} fill={$circle.fill} />
-			<text
-				x={$circle.x}
-				y={$circle.y}
-				font-family="JetBrains Mono"
-				font-size={$circle.r * 0.4}
-				text-anchor="middle"
-				dominant-baseline="middle"
-			>
-				{$circle.x.toFixed(0)}
-			</text>
-		</svg>
+		<!-- run arbitrary code -->
+		<div use:action={() => code.selectLines`2`}></div>
+		<div use:action={() => code.selectLines`3`}></div>
+		<div use:action={() => code.selectToken`double {double}`}></div>
+
+		<!-- code returns promise -->
+		<div
+			use:action={async () => {
+				await code.selectLines`*`
+				await code.update`
+					<script>
+						let count = $state(0)
+						let double = $derived(count * 2)
+					<\/script>
+
+					<button onclick={() => count++}>
+						{double}
+					</button>
+				`
+			}}
+		></div>
 	</Slide>
 
-	<Slide animate>
-		<p class="mt-8 text-6xl font-bold">🪄 Animotion</p>
-		<p class="mt-16 text-3xl">
-			Learn more by reading the
-			<a class="underline" href="https://animotion.pages.dev/docs" target="_blank">
-				Animotion docs
-			</a>.
-		</p>
+	<Slide class="h-full place-content-center place-items-center">
+		<h1 use:transition>Shuffle</h1>
+
+		<div class="mt-16 grid grid-cols-2 grid-rows-2 gap-4">
+			{#each boxes as box (box)}
+				<!-- custom CSS transition -->
+				<div
+					use:transition={{ transition: 'rotate' }}
+					class="grid h-[200px] w-[200px] place-content-center rounded-2xl bg-gray-100 font-semibold text-black shadow-2xl"
+				>
+					{box}
+				</div>
+			{/each}
+		</div>
+
+		<div use:transition={{ action: () => (boxes = [4, 3, 2, 1]) }}></div>
+		<div use:transition={{ action: () => (boxes = [2, 1, 4, 3]) }}></div>
+		<div use:transition={{ action: () => (boxes = [4, 3, 2, 1]) }}></div>
+		<div use:transition={{ action: () => (boxes = [1, 2, 3, 4]) }}></div>
+	</Slide>
+
+	<Slide class="h-full place-content-center place-items-center">
+		<div use:transition>
+			<p class="text-8xl font-bold drop-shadow-sm">🪄 Fin</p>
+		</div>
 	</Slide>
 </Presentation>
